requirements.txt
openai>=1.0.0
chromadb>=0.4.0
rich>=13.0.0
pyyaml>=6.0
numpy>=1.24.0
tiktoken>=0.5.0
